pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emodnet-viewer"
version = "1.1.0"
description = "Flask-based web application for visualizing EMODnet Seabed Habitats WMS layers with vector support"
readme = "CLAUDE.md"
authors = [
    {name = "EMODnet Viewer Development Team"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: Scientific/Engineering :: Visualization",
]
keywords = ["emodnet", "wms", "marine", "gis", "flask", "visualization", "geospatial", "vector", "gpkg"]
dependencies = [
    "Flask>=3.1.0",
    "requests>=2.32.0",
    "geopandas>=1.1.0",
    "Fiona>=1.10.0",
    "pyproj>=3.7.0",
    "Werkzeug>=3.1.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.2.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "flake8>=7.3.0",
    "black>=25.0.0",
    "isort>=6.0.0",
    "mypy>=1.18.0",
    "responses>=0.23.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "psutil>=5.9.0",
]
monitoring = [
    "psutil>=5.9.0",
    "redis>=4.6.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/your-org/emodnet-viewer"
Documentation = "https://github.com/your-org/emodnet-viewer/wiki"
Repository = "https://github.com/your-org/emodnet-viewer.git"
Issues = "https://github.com/your-org/emodnet-viewer/issues"

[project.scripts]
emodnet-dev = "scripts.dev_server:main"
emodnet-test = "scripts.run_tests:main"
emodnet-diagnostic = "scripts.wms_diagnostic:main"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["app.py", "config", "src", "tests", "scripts"]

# Coverage configuration
[tool.coverage.run]
source = ["app", "config", "src"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
ignore_errors = true
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "requests.*",
    "flask.*",
    "watchdog.*",
    "psutil.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=app",
    "--cov=config",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "wms: Tests that interact with WMS services",
    "slow: Tests that take a long time to run",
    "network: Tests that require network connectivity",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Setuptools configuration
[tool.setuptools.packages.find]
where = ["src"]
include = ["emodnet_viewer*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
emodnet_viewer = ["templates/*", "static/*"]